isGroupAdmin scope; separate graphql helpers from types
