+++ packages/hub/src/graphql/mutations/deleteMembership.ts
@@ -4,7 +4,7 @@ import {
   getMembership,
   getMyMembership,
   groupHasAdminsBesidesUser,
-} from "../types/Group";
+} from "../helpers/groupHelpers";
 
 builder.mutationField("deleteMembership", (t) =>
   t.withAuth({ signedIn: true }).fieldWithInput({
