+++ packages/hub/src/graphql/mutations/moveModel.ts
@@ -2,8 +2,9 @@ import { ZodError } from "zod";
 
 import { prisma } from "@/prisma";
 import { builder } from "../builder";
-import { Model, getWriteableModel } from "../types/Model";
-import { getWriteableOwnerBySlug } from "../types/Owner";
+import { Model } from "../types/Model";
+import { getWriteableModel } from "../helpers/modelHelpers";
+import { getWriteableOwnerBySlug } from "../helpers/ownerHelpers";
 import { validateSlug } from "../utils";
 import { NotFoundError } from "../errors/NotFoundError";
