+++ packages/hub/src/graphql/mutations/updateMembershipRole.ts
@@ -1,12 +1,11 @@
 import { prisma } from "@/prisma";
 import { builder } from "../builder";
+import { MembershipRoleType, UserGroupMembership } from "../types/Group";
 import {
-  MembershipRoleType,
-  UserGroupMembership,
   getMembership,
   getMyMembership,
   groupHasAdminsBesidesUser,
-} from "../types/Group";
+} from "../helpers/groupHelpers";
 
 builder.mutationField("updateMembershipRole", (t) =>
   t.withAuth({ signedIn: true }).fieldWithInput({
