+++ packages/hub/src/graphql/types/Model.ts
@@ -1,81 +1,9 @@
-import { Prisma, type Model as PrismaModel } from "@prisma/client";
-import { Session } from "next-auth";
+import { prismaConnectionHelpers } from "@pothos/plugin-prisma";
 
 import { builder } from "@/graphql/builder";
-import { prisma } from "@/prisma";
-import { prismaConnectionHelpers } from "@pothos/plugin-prisma";
-import { NotFoundError } from "../errors/NotFoundError";
 import { decodeGlobalIdWithTypename } from "../utils";
 import { ModelRevision, ModelRevisionConnection } from "./ModelRevision";
 import { Owner } from "./Owner";
-
-export function modelWhereHasAccess(
-  session: Session | null
-): Prisma.ModelWhereInput {
-  const orParts: Prisma.ModelWhereInput[] = [{ isPrivate: false }];
-  if (session) {
-    orParts.push({
-      owner: {
-        OR: [
-          {
-            user: { email: session.user.email },
-          },
-          {
-            group: {
-              memberships: {
-                some: {
-                  user: { email: session.user.email },
-                },
-              },
-            },
-          },
-        ],
-      },
-    });
-  }
-  return { OR: orParts };
-}
-
-export async function getWriteableModel({
-  session,
-  owner,
-  slug,
-}: {
-  session: Session;
-  owner: string;
-  slug: string;
-}): Promise<PrismaModel> {
-  // Note: `findUnique` would be safer, but then we won't be able to use nested queries
-  const model = await prisma.model.findFirst({
-    where: {
-      slug,
-      owner: {
-        slug: owner,
-        OR: [
-          {
-            user: { email: session.user.email },
-          },
-          {
-            group: {
-              memberships: {
-                some: {
-                  user: { email: session.user.email },
-                },
-              },
-            },
-          },
-        ],
-      },
-    },
-  });
-  if (!model) {
-    // FIXME - this will happen if permissions are not sufficient
-    // It would be better to throw a custom PermissionError
-    // (Note that we should throw PermissionError only if model is readable, but not writeable; otherwise it should still be "Can't find")
-    throw new NotFoundError("Can't find model");
-  }
-  return model;
-}
 
 export const Model = builder.prismaNode("Model", {
   id: { field: "id" },
