+++ packages/hub/src/graphql/types/User.ts
@@ -1,41 +1,13 @@
-import { type User as PrismaUser } from "@prisma/client";
-import { Session } from "next-auth";
-
-import { prisma } from "@/prisma";
-import { SignedInSession, builder } from "../builder";
+import { builder } from "../builder";
 import { GroupConnection, groupFromMembershipConnectionHelpers } from "./Group";
-import {
-  ModelConnection,
-  modelConnectionHelpers,
-  modelWhereHasAccess,
-} from "./Model";
+import { ModelConnection, modelConnectionHelpers } from "./Model";
+import { modelWhereHasAccess } from "../helpers/modelHelpers";
 import { Owner } from "./Owner";
 import {
   RelativeValuesDefinitionConnection,
   relativeValuesDefinitionConnectionHelpers,
 } from "./RelativeValuesDefinition";
-
-export function isSignedIn(
-  session: Session | null
-): session is SignedInSession {
-  return Boolean(session?.user.email);
-}
-
-export async function getSelf(session: SignedInSession) {
-  const user = await prisma.user.findUniqueOrThrow({
-    where: { email: session.user.email },
-  });
-  return user;
-}
-
-export async function isRootUser(user: PrismaUser) {
-  // see also: `isRootUser` auth scope in builder.ts
-  return Boolean(
-    user.email && user.emailVerified && ROOT_EMAILS.includes(user.email)
-  );
-}
-
-const ROOT_EMAILS = (process.env.ROOT_EMAILS ?? "").split(",");
+import { isRootUser } from "../helpers/userHelpers";
 
 export const User = builder.prismaNode("User", {
   id: { field: "id" },
